Move user container utilities to own file
