+++ Source Code/Ticket Management System/Ticket Management System.vcxproj
@@ -152,6 +152,7 @@
     <ClCompile Include="ticket_container_util.c" />
     <ClCompile Include="user.c" />
     <ClCompile Include="user_container.c" />
+    <ClCompile Include="user_container_util.c" />
     <ClCompile Include="vector.c" />
   </ItemGroup>
   <ItemGroup>
@@ -167,6 +168,7 @@
     <ClInclude Include="ticket_container_util.h" />
     <ClInclude Include="user.h" />
     <ClInclude Include="user_container.h" />
+    <ClInclude Include="user_container_util.h" />
     <ClInclude Include="vector.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
